moving occ back to env; resolving compilation issue
