--- projects/storefrontapp/src/environments/cds/cds.environment.dev.ts
@@ -1,10 +1,0 @@
-import { CdsEnvironment } from './cds.environment.types';
-
-export const cdsEnvironment: CdsEnvironment = {
-  enabled: true,
-  baseUrl: 'https://api.stage.context.cloud.sap',
-  profileTagJavaScriptUrl:
-    'https://tag.static.stage.context.cloud.sap/js/profile-tag.js',
-  profileTagConfigUrl:
-    'https://tag.static.stage.context.cloud.sap/config/dfbb97b0-f4d7-11e9-9c99-2125ab7968c6',
-};
