--- projects/storefrontapp/src/environments/cds/cds.environment.types.ts
@@ -1,6 +1,0 @@
-export interface CdsEnvironment {
-  enabled: boolean;
-  baseUrl?: string;
-  profileTagJavaScriptUrl?: string;
-  profileTagConfigUrl?: string;
-}
