+++ projects/storefrontapp/src/environments/environment.1905.ts
@@ -1,0 +1,7 @@
+import { Environment } from './models/environment.model';
+
+export const environment: Environment = {
+  production: true,
+  occBaseUrl: 'https://dev-com-17.accdemo.b2c.ydev.hybris.com:9002',
+  occApiPrefix: '/rest/v2/',
+};
