+++ projects/storefrontapp/src/environments/environment.2005.ts
@@ -1,0 +1,7 @@
+import { Environment } from './models/environment.model';
+
+export const environment: Environment = {
+  production: true,
+  occBaseUrl: 'https://dev-com-7.accdemo.b2c.ydev.hybris.com:9002',
+  occApiPrefix: '/occ/v2/',
+};
