--- projects/storefrontapp/src/environments/environment.prod.ts
@@ -1,5 +1,0 @@
-import { Environment } from './environment.types';
-
-export const environment: Environment = {
-  production: true,
-};
