+++ projects/storefrontapp/src/environments/models/environment.model.ts
@@ -1,3 +1,5 @@
 export interface Environment {
   production: boolean;
+  occBaseUrl: string;
+  occApiPrefix: string;
 }
