+++ projects/storefrontapp/src/environments/models/feature.environment.model.ts
@@ -1,0 +1,6 @@
+import { ModuleWithProviders } from '@angular/core';
+
+export interface FeatureEnvironment {
+  enabled: boolean;
+  feature?: ModuleWithProviders;
+}
