--- projects/storefrontapp/src/environments/occ/occ.environment.1905.ts
@@ -1,6 +1,0 @@
-import { OccEnvironment } from './occ.environment.types';
-
-export const occEnvironment: OccEnvironment = {
-  occBaseUrl: 'https://dev-com-17.accdemo.b2c.ydev.hybris.com:9002',
-  occApiPrefix: '/rest/v2/',
-};
