--- projects/storefrontapp/src/environments/occ/occ.environment.ts
@@ -1,6 +1,0 @@
-import { OccEnvironment } from './occ.environment.types';
-
-export const occEnvironment: OccEnvironment = {
-  occBaseUrl: '',
-  occApiPrefix: '/rest/v2/',
-};
