--- projects/storefrontapp/src/environments/occ/occ.environment.types.ts
@@ -1,4 +1,0 @@
-export interface OccEnvironment {
-  occBaseUrl: string;
-  occApiPrefix: string;
-}
